LONG THANH INT

+++ SWP391_BE/src/main/java/com/swp391/hairsalon/controller/AppointmentController.java
@@ -271,7 +271,7 @@ public class AppointmentController {
 
     private List<SalonService> getSalonServicesById(AppointmentRequestDTO appointmentRequest) {
         List<SalonService> services = new ArrayList<>();
-        for (Long serviceId : appointmentRequest.getServiceId()) {
+        for (int serviceId : appointmentRequest.getServiceId()) {
             SalonService service = iSalonServiceService.getServiceById(serviceId);
             if (service != null) {
                 services.add(service);
@@ -282,7 +282,7 @@ public class AppointmentController {
 
     private int getLoyalPoint(AppointmentRequestDTO appointmentRequest){
         int loyalPoint=0;
-        for (Long serviceId : appointmentRequest.getServiceId()) {
+        for (int serviceId : appointmentRequest.getServiceId()) {
             SalonService service = iSalonServiceService.getServiceById(serviceId);
             if (service != null) {
                 loyalPoint = loyalPoint +  (int) Math.ceil(service.getServicePrice());
